Account for floats

+++ btree_lowest_sum.py
@@ -20,7 +20,7 @@ def level_with_min_sum(root):
             continue
         elif digits:  #  add digits to whatever level we are on
             sums[level] = 0 if level not in sums else sums[level]
-            sums[level] += int(digits)
+            sums[level] += float(digits)
         if char in "[]":  # increment or decrement current level
             level += (1 if char == "[" else -1)
         digits = ""
